backends: http: Add http.factory()
Preparing for multiple connections support, we need a backend factory
creating multiple connections to same remote server. The factory is also
useful in the tests for cleaning up repeating code for opening http
backend to the test http server.

The factory is simplistic; we may optimize it later to perform one
OPTIONS request for the entire pool instead of repeating this for every
connection.

Change-Id: I063da29175e0a5efa773ffdf010f358b9573d8ce
Signed-off-by: Nir Soffer <nsoffer@redhat.com>