backends: memory: Support external backing
Change memory backend to accept external io.BytesIO() instead of copying
the data to internal io.BytesIO(). This make it possible to create
multiple readers sharing the same backing, and testing backing contents
after passing memory.factory() to io.copy().

Since the backing file may be shared, the backend does not not own it
and does not close it. Instead the backend maintain closed state to
behave like other backends.

Change-Id: Icb82df381bacadec93a7f610cfdba99f80a005ed
Signed-off-by: Nir Soffer <nsoffer@redhat.com>