client: Support multiple connections
Reimplement io.copy() using multiple connections. Initial tests show 35%
improvement but more testing is needed with real environment.

Every worker thread opens a connection to the source and destination
backends, and process I/O requests submitted to a work queue by the main
thread.

Multiple connections can work only when accessing imageio daemon
reporting the number of allowed readers and writers. When accessing old
imageio daemon we use single connection.

Client test assuming single thread was changed so it does not assume the
order of the requests, which is non-deterministic with multiple
connections. The test is marked as xfail on Jenkins since we get
unexpected results on the local file system there. We need to use user
storage for predictable results.

Change-Id: Id1ffad521ca5349da7cace4d49d261fde081f48d
Bug-Url: https://bugzilla.redhat.com/1591439
Signed-off-by: Nir Soffer <nsoffer@redhat.com>