http: Setup 60 seconds connection timeout
The default timeout (None) results in waiting for new request for more
than 300 seconds if the client keeps the connection open. This happens
when uploading or downloading with engine 4.4.

This also shorten the time we wait when receiving or sending data over
the connection. Timeout means the client is not healthy and the only way
to handle this is to close the connection.

Engine 4.4 need to be fixed to close the connection when done. Until it
is fixed, we will see a warning when the connection is closed after a
timeout.

Older engine did not have this problem since old proxy did not support
keep-alive connections and closed the connection after every request.

Change-Id: Ic5d9232ea73e6668cf0488a745755c0bd8dfa95b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>