images: Allow stupid client to close connection
When downloading disks via ovirt-engine UI, the UI does not have any
control over the connection. The browser keeps the connection open for a
while, optimizing future request to this domain. This leads to failure
to deactivate a volume after download.

The problem was hidden in the past since old proxy did not support
keep-alive connections and was closing the connection after every
request.

Add a close=y query argument, allowing stupid clients to specify the
equivalent of "Connection: close" header in a get request.

ovirt-engine is expected to download images now using:

    https://proxy-address:54323/images/ticket-id?close=y

Documentation and tests will be added later.

Change-Id: Iaebbc4ad7d84b55d296983f612a80c76372acae7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>