spec: Add ovirt-imageio-client package
The client package adds the client.py and ui.py modules that are the
only public part of the ovirt_imageio package.

This package depends on qemu-img, so we don't need to require it in
ovirt-imageio-common. qemu-img is needed only for the client library.

This will allow requiring newer qemu-img from the advanced virt module,
which may not be available on ovirt-engine host.

Change-Id: Ide69388585580f31241d8b375e72fecfea288c91
Signed-off-by: Nir Soffer <nsoffer@redhat.com>