logging: Add default stderr handler
New unified configuration is very restrictive, allowing only sections
and options defined in config.py. This makes it hard to override the
logger to log to stderr in engine development config.

Add a default [handler_stderr] section, so we can use this template
configuration:

    [handlers]
    keys = {handler}

    [logger_root]
    handlers = {handler}
    level = {level}

With this template it is easy to create a production or development
logging configuration dynamically when running engine-setup.

It would nice if we did not have to override handlers, but looks like
all handlers are created even if no logger use them.

Change-Id: I5d78e4c2fcbf9b4e4c1f7a6333f10132c20e9f82
Signed-off-by: Nir Soffer <nsoffer@redhat.com>