qemu_nbd: Add backing_chain argument
When reading qcow2 image qemu-nbd reads data from the entire backing
chain. But in some cases we like to read data only from a single image.
One use case is to upload chain of snapshots back to storage.

Add backing_chain=True argument to qemu_nbd.run() and qemu_nbd.Server().
When used, we specify {"backing": null} to disable the backing file, and
qemu-nbd reports unallocated areas as holes.

We use now "json:" filename argument instead of the actual filename.
This gives us lot of control on the way the image will be opened, and is
more powerful and easy to use compared with --image-opts.

For normal image, the command will be now:

    qemu-nbd ... 'json:{"driver": "qcow2", "file": {"driver":"file","filename":"file.qcow2"}}'

When using backing_chain=False:

    qemu-nbd ... 'json:{"driver": "qcow2", "file": {"driver":"file","filename":"file.qcow2"},"backing": null}'

Change-Id: I6d7bc8034ee8e151e427e29475413b880f08784c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>