extents: Report holes in zero extents
NBD protocol reports if an extents is a hole. This means that the area
is unallocated.

When getting extents for qcow2 image, areas which are not allocated in
the image are reported as holes. When reading from unallocated areas
qemu reads data from the backing chain.

Note that when getting extents from raw images, unallocated areas are
not reported as holes. This is a limitation of qemu nbd server.

We are interested in holes to be able to download and upload single
qcow2 snapshot correctly, so reconstructing the chain will restore the
same data as seen by the guest.

Change-Id: I0351e15ca6f87ff96a0616fad8bfc4ec80f29906
Signed-off-by: Nir Soffer <nsoffer@redhat.com>