qemu_nbd: Export images from ova
qemu_nbd module supports now exporting images from OVA file[1]. From our
point of view a OVA is a tar file with one or more disks. The user should
find the disks in the OVA file, and get their format, offset in the tar
file, and size of disk in the tar.

With the information about the disk, we can export the image using:

    with qemu_nbd.run(
            "vm.ova",
            "qcow2",
            "/tmp/nbd.sock",
            offset=1536,
            size=676659200):
        # Access image using /tmp/nbd.sock here...

This exports the qcow2 image, which may be compressed as raw data,
reading directly from the tar file.

We use 3 different 'json:{...}' filename now:

1. Exporting qcow2 disk from a range in tar file

    'json:{"driver": "qcow2",
           "file": {"driver": "raw",
                    "offset": 1536,
                    "size": 676659200,
                    "file": {"driver": "file",
                             "filename": "vm.ova"}}}'

This is the typical format we expect to handle.

2. Exporting raw disk from a range in tar file

    'json:{"driver": "raw",
           "offset": 1536,
           "size": 676659200,
           "file": {"driver": "file",
                    "filename": "vm.ova"}}'

I'm not sure this format is useful, but we can support it.

3. Exporting entire image

    'json:{"driver": "qcow2",
           "file": {"driver": "file",
                    "filename": "disk.qcow2"}}'

This was the only format we supported before this change.

The size argument is not needed when using qcow2 format but required
when exporting raw image, otherwise you export the entire image from the
given offset.

Compressed tar file is not supported since OVA does not supports this
and it would not be efficient to process.

Compressed raw disks are also not supported since qemu-nbd does not
support raw compressed images.

[1] https://www.dmtf.org/sites/default/files/standards/documents/DSP0243_1.1.0.pdf

Change-Id: I8eb368e9a6aec05decb7a318a365fc46112ef882
Bug-Url: https://bugzilla.redhat.com/1849981
Signed-off-by: Nir Soffer <nsoffer@redhat.com>