auth: Enable ticket cancellation
Previously we removed tickets without checking the ticket status. Now
removing a ticket will cancel the ticket and wait until the ticket is
unused.

If the ticket is still used after the remove timeout, the request will
fail with "409 Conflict". The caller need to poll the ticket status
until the number of connections reaches zero, and then remove the ticket
again. A simpler option is to retry removing the ticket until the
operation succeeds.

When a ticket is canceled, authorizing a ticket will fail. Add tests for
the Authorizer that previously was test only by the functional tests.

Ticket remove timeout can be modified using new config option:

    [control]
    remove_timeout = 60

Testing canceling upload using 4 connections and local storage show that
cancelling takes 30-650 milliseconds.

More tests are needed:

- functional tests for canceling tickets an monitoring number of
  connections during cancellation.
- functional tests for cancel timeout
- stress test for cancelling tickets

Change-Id: I0c9462e1fc927a8e4a572fefc62b4d6adeccb3ab
Bug-Url: https://bugzilla.redhat.com/1524184
Signed-off-by: Nir Soffer <nsoffer@redhat.com>