nbd: More helpful reply error
Previously when qemu-nbd failed we got an error like:

    [Error 28] bytearray(b'writing to file failed')

This error is not helpful since qemu-nbd message is too generic, and the
actual error (ENOSPC) is hidden behind the error number. To make this
worse we did not decode the error message bytearray.

Change ReplyError to treat message as an error message, and code as the
reason. Now we show the same error in a more helpful way:

    Writing to file failed: [Error 28] No space left on device

When handling a simple reply error when we have no message we use
"Simple reply failed":

    Simple reply failed: [Error 1] Operation not permitted

qemu-nbd always send a message, but the message is optional. If the
server did not send a message, show a generic error message:

    Server error: [Error 5] Input/output error

Change-Id: Ib05a5f6cb011321cf860e78c99fe1a8f9a8e655d
Bug-Url: https://bugzilla.redhat.com/1854550
Signed-off-by: Nir Soffer <nsoffer@redhat.com>