ops: Rename confusing Send and Receive
Using Send and Receive was a bad choice. We use the terms READ, WRITE,
ZERO, and FLUSH in the logs, and having Send and Receive operations does
not make sense.

The only place when we keep the old name is the directio module which is
backward compatibility hack for vdsm, using this internal API.

This make the stats somewhat more confusing using "read.read" and
"write.write" but at least this is consistent with "zero.zero".

Here are example stats with the new names:

Upload:

[connection 1 ops, 3.199483 s]
[dispatch 55 ops, 3.093170 s]
[write 23 ops, 3.018835 s, 333.69 MiB, 110.54 MiB/s]
[write.read 180 ops, 1.437334 s, 333.69 MiB, 232.16 MiB/s]
[write.write 180 ops, 1.567051 s, 333.69 MiB, 212.94 MiB/s]
[zero 30 ops, 0.045386 s, 1019.12 MiB, 21.93 GiB/s]
[zero.zero 30 ops, 0.044577 s, 1019.12 MiB, 22.33 GiB/s]
[flush 1 ops, 0.000543 s]

Download:

[connection 1 ops, 2.361870 s]
[dispatch 49 ops, 2.177001 s]
[read 49 ops, 2.161489 s, 343.50 MiB, 158.92 MiB/s]
[read.read 203 ops, 1.577760 s, 343.50 MiB, 217.71 MiB/s]
[read.write 203 ops, 0.563805 s, 343.50 MiB, 609.25 MiB/s]

Change-Id: I4b988d4781f7b29d46a0ad44a411488f7723be35
Signed-off-by: Nir Soffer <nsoffer@redhat.com>