client: Fix hole handling
Previously io.copy() supported zero flag for controlling zeroing of zero
and hole extents. When zero=True, we zeroed both zero and hole extents.
Otherwise we skipped both zero and hole extents. This is correct when
uploading to single image, or when downloading a single image.

When copying to image with a backing file, zero extents in destination
image must hide data from the backing chain, and holes must not be
allocated so they expose data from the backing chain.

To allow this behaviour, add a hole=True argument to io.copy(). When
hole=False, skip holes to keep them unallocated in the destination
image, exposing data from the backing chain.

This complexity is not relevant to dirty extents, so the copy loop was
separate again to _copy_dirty() and _copy_data().

Change-Id: Idd5f7bf4844f6b939f5f5330c1e994236eedbd83
Bug-Url: https://bugzilla.redhat.com/1847090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>