client: Support dirty extents
client.extents() supports now bitmap argument, reporting the dirty
extents in this bitmap.

This works only for qcow2 image, and the specified bitmap must exists in
the image.

This can be used for testing code merging bitmaps, or to analyze image
with dirty bitmaps on storage.

This works also for images inside OVA archive, but since bitmaps are not
included in downloaded images, finding images with bitmaps in OVA is
very unlikely.

Since qemu-nbd report only dirty bitmap from the top image, the dirty
extents are correct only for the top image.

Change-Id: I6c3f76e737263d60cd633540cba04d4609bcca1e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>